Task 04 (phase 5): extract :core:designsystem (theme + composables)
Move ui/theme/* and the reusable composables (ChartView, ExpandableText,
LinkText, RatingView, RoundedTabSection, SnackbarMessage, TextFieldWithError)
into a new Compose :core:designsystem module, along with the device-only content
drawables (ic_arrow_back, ic_selected_heart, ic_unselected_heart, ic_star*).

Rewrite theme/composable imports across features + app, and point the two
:feature:device files at com.timhome.core.designsystem.R for the moved drawables.
Launcher mipmaps / app themes / strings stay in :core for now (Phase 8 -> :app).

designsystem depends on :core:{common,model,database} — the database dep exists
only because ChartView renders CarbonDioxideEntity directly (pre-existing UI/DB
coupling, noted for a later model refactor). Navigation.kt transitions stay in
:core for now (move to :core:ui in phase 6).

Builds green: assembleDebug + detektDebug + ktlintCheck.
