+++ app/build.gradle.kts
@@ -72,6 +72,7 @@ android {
 dependencies {
     implementation(project(":core"))
     implementation(project(":core:common"))
+    implementation(project(":core:designsystem"))
     implementation(project(":core:data"))
     implementation(project(":core:database"))
     implementation(project(":base"))
