+++ app/src/main/java/com/timhome/modularizationtest/ui/MainActivity.kt
@@ -71,7 +71,7 @@ import androidx.navigation.compose.rememberNavController
 import com.timhome.auth.ui.signin.navigation.signInRoute
 import com.timhome.base.DaggerBaseComponent
 import com.timhome.core.coreComponent
-import com.timhome.core.ui.theme.HomeTheme
+import com.timhome.core.designsystem.theme.HomeTheme
 import com.timhome.core.common.NavigationDispatcher
 import com.timhome.device.ui.navigation.deviceRoute
 import com.timhome.home.ui.navigation.homeRoute
