+++ core/designsystem/src/main/java/com/timhome/core/designsystem/theme/Shape.kt
@@ -1,4 +1,4 @@
-package com.timhome.core.ui.theme
+package com.timhome.core.designsystem.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.MaterialTheme
