+++ core/designsystem/src/main/res/drawable/ic_selected_heart.xml
