+++ core/designsystem/src/main/res/drawable/ic_unselected_heart.xml
