+++ core/designsystem/src/main/res/mipmap-hdpi/ic_star.png
