+++ core/designsystem/src/main/res/mipmap-hdpi/ic_star_filled.png
