+++ core/designsystem/src/main/res/mipmap-mdpi/ic_star_filled.png
