+++ core/designsystem/src/main/res/mipmap-xxhdpi/ic_star_filled.png
