+++ core/designsystem/src/main/res/mipmap-xxxhdpi/ic_star_filled.png
