+++ feature/auth/src/main/java/com/timhome/auth/ui/signin/composable/AuthTabSection.kt
@@ -18,8 +18,8 @@ import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.timhome.core.ui.RoundedTabSection
-import com.timhome.core.ui.theme.AuthTabSectionBackgroundColor
+import com.timhome.core.designsystem.RoundedTabSection
+import com.timhome.core.designsystem.theme.AuthTabSectionBackgroundColor
 import com.timhome.core.common.OnEnterText
 import com.timhome.core.utils.rememberImeState
 import kotlinx.collections.immutable.PersistentList
