+++ feature/device/src/main/java/com/timhome/device/ui/buyModule/BuyModuleScreen.kt
@@ -36,8 +36,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.timhome.core.ui.theme.AuthTabSectionBackgroundColor
-import com.timhome.core.ui.theme.HomeTheme
+import com.timhome.core.designsystem.theme.AuthTabSectionBackgroundColor
+import com.timhome.core.designsystem.theme.HomeTheme
 import com.timhome.core.utils.viewmodel.ViewModelFactory
 import com.timhome.device.R
 import com.timhome.device.data.model.ModuleModel
