+++ feature/device/src/main/java/com/timhome/device/ui/device/DeviceDetailScreen.kt
@@ -19,8 +19,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.timhome.core.ui.theme.HomeTheme
-import com.timhome.core.ui.theme.cornerRoundedShapes
+import com.timhome.core.designsystem.theme.HomeTheme
+import com.timhome.core.designsystem.theme.cornerRoundedShapes
 import com.timhome.core.utils.viewmodel.ViewModelFactory
 import com.timhome.device.ui.device.composables.LazyColumnWithParallax
