+++ feature/device/src/main/java/com/timhome/device/ui/listDevice/DeviceListScreen.kt
@@ -12,7 +12,7 @@ import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.timhome.core.ui.theme.HomeTheme
+import com.timhome.core.designsystem.theme.HomeTheme
 import com.timhome.core.common.trace
 import com.timhome.core.utils.viewmodel.ViewModelFactory
 import com.timhome.device.domain.models.DeviceModel
