+++ feature/home/src/main/java/com/timhome/home/custom/Co2Indicator.kt
@@ -32,12 +32,12 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.timhome.core.ui.theme.IndicatorCO2Acceptable
-import com.timhome.core.ui.theme.IndicatorCO2Bed
-import com.timhome.core.ui.theme.IndicatorCO2Danger
-import com.timhome.core.ui.theme.IndicatorCO2Good
-import com.timhome.core.ui.theme.IndicatorCO2LowDanger
-import com.timhome.core.ui.theme.IndicatorCO2Uncomfortable
+import com.timhome.core.designsystem.theme.IndicatorCO2Acceptable
+import com.timhome.core.designsystem.theme.IndicatorCO2Bed
+import com.timhome.core.designsystem.theme.IndicatorCO2Danger
+import com.timhome.core.designsystem.theme.IndicatorCO2Good
+import com.timhome.core.designsystem.theme.IndicatorCO2LowDanger
+import com.timhome.core.designsystem.theme.IndicatorCO2Uncomfortable
 import com.timhome.core.common.Constant.INDICATOR_CO2_ACCEPTABLE_VALUE
 import com.timhome.core.common.Constant.INDICATOR_CO2_BED_LEVEL
 import com.timhome.core.common.Constant.INDICATOR_CO2_GOOD_LEVEL
