+++ feature/home/src/main/java/com/timhome/home/ui/HomeScreen.kt
@@ -32,10 +32,10 @@ import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.timhome.core.database.entity.CarbonDioxideEntity
-import com.timhome.core.ui.ChartView
-import com.timhome.core.ui.theme.CO2BackgroundColor
-import com.timhome.core.ui.theme.DataCreatedItemColor
-import com.timhome.core.ui.theme.HomeTheme
+import com.timhome.core.designsystem.ChartView
+import com.timhome.core.designsystem.theme.CO2BackgroundColor
+import com.timhome.core.designsystem.theme.DataCreatedItemColor
+import com.timhome.core.designsystem.theme.HomeTheme
 import com.timhome.core.utils.viewmodel.ViewModelFactory
 import com.timhome.home.R
 import com.timhome.home.custom.Co2Indicator
