+++ feature/settings/src/main/java/com/timhome/settings/ui/SettingsScreen.kt
@@ -26,8 +26,8 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.timhome.core.ui.theme.HomeTheme
-import com.timhome.core.ui.theme.TextFieldBackgroundColor
+import com.timhome.core.designsystem.theme.HomeTheme
+import com.timhome.core.designsystem.theme.TextFieldBackgroundColor
 import com.timhome.core.utils.viewmodel.ViewModelFactory
 import com.timhome.settings.R
 import com.timhome.settings.ui.composables.SwitchWithText
