+++ settings.gradle.kts
@@ -29,6 +29,7 @@ include(":core:network")
 include(":core:database")
 include(":core:datastore")
 include(":core:data")
+include(":core:designsystem")
 include(":benchmark")
 include(":mock")
 include(":base")
